refactor(contracts): separate capabilities from compose schemas
