+++ turbo/apps/api/src/signals/auth/auth-context.ts
@@ -9,7 +9,7 @@ import {
   verifyZeroToken,
 } from "./tokens";
 import { clerkSessionAuth$ } from "./clerk-session";
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 import { AuthContext, CliAuth, ZeroAuthContext } from "../../types/auth";
 import {
   cliTokenRecord,
