+++ turbo/apps/api/src/signals/routes/__tests__/chat-threads-rename.test.ts
@@ -3,7 +3,7 @@ import {
   chatThreadMetadataContract,
   chatThreadRenameContract,
 } from "@okouai/api-contracts/contracts/chat-threads";
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 import { DEFAULT_ORG_MODEL_POLICY_DEFAULT_MODEL } from "@okouai/api-contracts/contracts/model-providers";
 import { createStore } from "ccstate";
 import { describe, expect, it } from "vitest";
