+++ turbo/apps/api/src/signals/routes/__tests__/chat-threads-video-model.test.ts
@@ -5,7 +5,7 @@ import {
   chatThreadsContract,
   chatThreadVideoModelContract,
 } from "@okouai/api-contracts/contracts/chat-threads";
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 import { createStore } from "ccstate";
 import { describe, expect, it } from "vitest";
