+++ turbo/apps/api/src/signals/routes/__tests__/image-recognition.test.ts
@@ -5,7 +5,7 @@ import {
   IMAGE_RECOGNITION_MAX_FILE_BYTES,
   imageRecognitionContract,
 } from "@okouai/api-contracts/contracts/image-recognition";
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 import { usageRecordContract } from "@okouai/api-contracts/contracts/usage-record";
 import { HttpResponse, http } from "msw";
 import { onTestFinished } from "vitest";
