+++ turbo/apps/api/src/signals/routes/__tests__/uploads-complete.test.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from "node:crypto";
 
 import { describe, expect, it } from "vitest";
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 
 import { testContext } from "../../../__tests__/test-context";
 import { now } from "../../../lib/time";
