+++ turbo/apps/api/src/types/auth.ts
@@ -1,4 +1,4 @@
-import type { ZeroCapability } from "@okouai/api-contracts/contracts/composes";
+import type { ZeroCapability } from "@okouai/api-contracts/contracts/capabilities";
 
 export type ApiOrgRole = "admin" | "member";
